[batch] prevent test from evaluating server.py
